--- src/franc/index.ts
@@ -1,11 +1,0 @@
-import Money from '../money';
-
-export default class Franc extends Money {
-  constructor(amount: number) {
-    super(amount);
-  }
-
-  times(multiplier: number): Franc {
-    return new Franc(this.amount * multiplier);
-  }
-}
