+++ src/money/index.ts
@@ -31,4 +31,14 @@ export class Dollar extends Money {
   }
 }
 
+export class Franc extends Money {
+  constructor(amount: number) {
+    super(amount);
+  }
+
+  times(multiplier: number): Franc {
+    return new Franc(this.amount * multiplier);
+  }
+}
+
 export default Money;
